+++ CHANGELOG.md
@@ -47,6 +47,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- `new ImageSource()` will now log a warning if an image type isn't fully supported. ([#1855](https://github.com/excaliburjs/Excalibur/issues/1855))
 - `Timer.start()` to explicitly start timers, and `Timer.stop()` to stop timers and "rewind" them.
 - `Timer.timeToNextAction` will return the milliseconds until the next action callback
 - `Timer.timeElapsedTowardNextAction` will return the milliseconds counted towards the next action callback